Improved the connectivity between clue generation and solvability to fix GUI issue
